Added GUI class to handle the rendering of emulator graphics. Added ROM class to contain file details and organize file related functions.
Catch2 removed as not necessary yet since most of the interfaces for classes
are private.

A work-in-progress GUI for the debugger/logger written in Python 3.7.3 using
tkinter is being worked on, mostly as a visual aid for observing details of
the emulator. This is to avoid having to constantly open files to debug,
eventually debugger will be able to read register data for CPU and PPU
and set breakpoints if deemed necessary.